+++ Assets/Scripts/Entity/MarbleCommander.cs
@@ -6,11 +6,11 @@ using UnityTimer;
 
 public class MarbleCommander : SteveCommander
 {
-    public MarbleCommander(int uid, Player player) : base(uid, player)
+    public MarbleCommander(long uid, Player player) : base(uid, player)
     {
     }
 
-    public MarbleCommander(int uid, Player player, Color color) : base(uid, player, color)
+    public MarbleCommander(long uid, Player player, Color color) : base(uid, player, color)
     {
     }
 }
